+++ javaslang/src/main/java/javaslang/collection/TreeSet.java
@@ -5,11 +5,17 @@
  */
 package javaslang.collection;
 
-import javaslang.*;
+import javaslang.Kind1;
+import javaslang.Tuple;
+import javaslang.Tuple2;
+import javaslang.Tuple3;
 import javaslang.control.Option;
 
 import java.io.Serializable;
-import java.util.*;
+import java.util.ArrayList;
+import java.util.Comparator;
+import java.util.NoSuchElementException;
+import java.util.Objects;
 import java.util.function.*;
 import java.util.stream.Collector;
 
@@ -560,7 +566,7 @@ public final class TreeSet<T> implements Kind1<TreeSet<?>, T>, SortedSet<T>, Ser
     }
 
     @Override
-    public TreeSet<T> drop(long n) {
+    public TreeSet<T> drop(int n) {
         if (n <= 0) {
             return this;
         } else if (n >= length()) {
@@ -571,7 +577,7 @@ public final class TreeSet<T> implements Kind1<TreeSet<?>, T>, SortedSet<T>, Ser
     }
 
     @Override
-    public TreeSet<T> dropRight(long n) {
+    public TreeSet<T> dropRight(int n) {
         if (n <= 0) {
             return this;
         } else if (n >= length()) {
@@ -625,7 +631,7 @@ public final class TreeSet<T> implements Kind1<TreeSet<?>, T>, SortedSet<T>, Ser
     }
 
     @Override
-    public Iterator<TreeSet<T>> grouped(long size) {
+    public Iterator<TreeSet<T>> grouped(int size) {
         return sliding(size, size);
     }
 
@@ -796,12 +802,12 @@ public final class TreeSet<T> implements Kind1<TreeSet<?>, T>, SortedSet<T>, Ser
     }
 
     @Override
-    public Iterator<TreeSet<T>> sliding(long size) {
+    public Iterator<TreeSet<T>> sliding(int size) {
         return sliding(size, 1);
     }
 
     @Override
-    public Iterator<TreeSet<T>> sliding(long size, long step) {
+    public Iterator<TreeSet<T>> sliding(int size, int step) {
         return iterator().sliding(size, step).map(seq -> TreeSet.ofAll(tree.comparator(), seq));
     }
 
@@ -827,7 +833,7 @@ public final class TreeSet<T> implements Kind1<TreeSet<?>, T>, SortedSet<T>, Ser
     }
 
     @Override
-    public TreeSet<T> take(long n) {
+    public TreeSet<T> take(int n) {
         if (n <= 0) {
             return empty(tree.comparator());
         } else if (n >= length()) {
@@ -838,7 +844,7 @@ public final class TreeSet<T> implements Kind1<TreeSet<?>, T>, SortedSet<T>, Ser
     }
 
     @Override
-    public TreeSet<T> takeRight(long n) {
+    public TreeSet<T> takeRight(int n) {
         if (n <= 0) {
             return empty(tree.comparator());
         } else if (n >= length()) {
@@ -930,14 +936,14 @@ public final class TreeSet<T> implements Kind1<TreeSet<?>, T>, SortedSet<T>, Ser
     }
 
     @Override
-    public TreeSet<Tuple2<T, Long>> zipWithIndex() {
+    public TreeSet<Tuple2<T, Integer>> zipWithIndex() {
         final Comparator<? super T> component1Comparator = tree.comparator();
-        final Comparator<Tuple2<T, Long>> tuple2Comparator = (t1, t2) -> component1Comparator.compare(t1._1, t2._1);
+        final Comparator<Tuple2<T, Integer>> tuple2Comparator = (t1, t2) -> component1Comparator.compare(t1._1, t2._1);
         return TreeSet.ofAll(tuple2Comparator, iterator().zipWithIndex());
     }
 
     @Override
-    public <U> SortedSet<U> zipWithIndex(BiFunction<? super T, ? super Long, ? extends U> mapper) {
+    public <U> SortedSet<U> zipWithIndex(BiFunction<? super T, ? super Integer, ? extends U> mapper) {
         return TreeSet.ofAll(naturalComparator(), iterator().zipWithIndex(mapper));
     }
